+++ sdk/recoveryservicesbackup/azure-resourcemanager-recoveryservicesbackup/src/main/java/com/azure/resourcemanager/recoveryservicesbackup/models/AzureVmWorkloadSapAseDatabaseProtectedItem.java
@@ -102,6 +102,13 @@ public final class AzureVmWorkloadSapAseDatabaseProtectedItem extends AzureVmWor
 
     /** {@inheritDoc} */
     @Override
+    public AzureVmWorkloadSapAseDatabaseProtectedItem withNodesList(List<DistributedNodesInfo> nodesList) {
+        super.withNodesList(nodesList);
+        return this;
+    }
+
+    /** {@inheritDoc} */
+    @Override
     public AzureVmWorkloadSapAseDatabaseProtectedItem withContainerName(String containerName) {
         super.withContainerName(containerName);
         return this;
@@ -205,8 +212,9 @@ public final class AzureVmWorkloadSapAseDatabaseProtectedItem extends AzureVmWor
 
     /** {@inheritDoc} */
     @Override
-    public AzureVmWorkloadSapAseDatabaseProtectedItem withSoftDeleteRetentionPeriod(Integer softDeleteRetentionPeriod) {
-        super.withSoftDeleteRetentionPeriod(softDeleteRetentionPeriod);
+    public AzureVmWorkloadSapAseDatabaseProtectedItem withSoftDeleteRetentionPeriodInDays(
+        Integer softDeleteRetentionPeriodInDays) {
+        super.withSoftDeleteRetentionPeriodInDays(softDeleteRetentionPeriodInDays);
         return this;
     }