+++ sdk/recoveryservicesbackup/azure-resourcemanager-recoveryservicesbackup/src/main/java/com/azure/resourcemanager/recoveryservicesbackup/models/AzureVmWorkloadSqlDatabaseProtectedItem.java
@@ -101,6 +101,13 @@ public final class AzureVmWorkloadSqlDatabaseProtectedItem extends AzureVmWorklo
 
     /** {@inheritDoc} */
     @Override
+    public AzureVmWorkloadSqlDatabaseProtectedItem withNodesList(List<DistributedNodesInfo> nodesList) {
+        super.withNodesList(nodesList);
+        return this;
+    }
+
+    /** {@inheritDoc} */
+    @Override
     public AzureVmWorkloadSqlDatabaseProtectedItem withContainerName(String containerName) {
         super.withContainerName(containerName);
         return this;
@@ -202,8 +209,9 @@ public final class AzureVmWorkloadSqlDatabaseProtectedItem extends AzureVmWorklo
 
     /** {@inheritDoc} */
     @Override
-    public AzureVmWorkloadSqlDatabaseProtectedItem withSoftDeleteRetentionPeriod(Integer softDeleteRetentionPeriod) {
-        super.withSoftDeleteRetentionPeriod(softDeleteRetentionPeriod);
+    public AzureVmWorkloadSqlDatabaseProtectedItem withSoftDeleteRetentionPeriodInDays(
+        Integer softDeleteRetentionPeriodInDays) {
+        super.withSoftDeleteRetentionPeriodInDays(softDeleteRetentionPeriodInDays);
         return this;
     }