+++ sdk/recoveryservicesbackup/azure-resourcemanager-recoveryservicesbackup/src/main/java/com/azure/resourcemanager/recoveryservicesbackup/models/DistributedNodesInfo.java
@@ -29,6 +29,12 @@ public final class DistributedNodesInfo {
     @JsonProperty(value = "errorDetail")
     private ErrorDetail errorDetail;
 
+    /*
+     * ARM resource id of the node
+     */
+    @JsonProperty(value = "sourceResourceId")
+    private String sourceResourceId;
+
     /** Creates an instance of DistributedNodesInfo class. */
     public DistributedNodesInfo() {
     }
@@ -90,6 +96,26 @@ public final class DistributedNodesInfo {
      */
     public DistributedNodesInfo withErrorDetail(ErrorDetail errorDetail) {
         this.errorDetail = errorDetail;
+        return this;
+    }
+
+    /**
+     * Get the sourceResourceId property: ARM resource id of the node.
+     *
+     * @return the sourceResourceId value.
+     */
+    public String sourceResourceId() {
+        return this.sourceResourceId;
+    }
+
+    /**
+     * Set the sourceResourceId property: ARM resource id of the node.
+     *
+     * @param sourceResourceId the sourceResourceId value to set.
+     * @return the DistributedNodesInfo object itself.
+     */
+    public DistributedNodesInfo withSourceResourceId(String sourceResourceId) {
+        this.sourceResourceId = sourceResourceId;
         return this;
     }